backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.133.1",
    "huggingface-hub>=1.5.0",
    "numpy>=2.4.2",
    "onnxruntime>=1.24.2",
    "opencv-python>=4.13.0.92",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "sam2>=1.1.0",
    "segment-anything>=1.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "uvicorn[standard]>=0.41.0",
]
